requirements.txt
pandas
requests
psycopg2-binary
schedule
python-dotenv
